added test for part2
