+++ app/views/campaign_finances/index.html.erb
@@ -11,7 +11,6 @@
     <%= form.select :category, options_for_select([
       ['Candidate Loan', 'candidate-loan'],
       ['Contribution Total', 'contribution-total'],
-      # ... other categories
     ]) %>
   </div>
